requirements.txt
Authomatic==0.1.0.post1
Django==1.8.12
Pillow==3.1.1
PyJWT==1.4.0
argparse==1.2.1
beautifulsoup4==4.4.1
dj-database-url==0.4.0
django-appconf==1.0.1
django-googledrive-storage==1.1.3
django-location-field==1.6.1
feedparser==5.2.1
freeze==0.8.0
google-api-python-client==1.5.0
googlemaps==2.4.3
gunicorn==19.4.5
httplib2==0.9.2
httpretty==0.8.3
mock==1.0.1
nose==1.3.7
oauth2==1.9.0.post1
oauth2client==2.0.1
oauthlib==1.0.3
psycopg2==2.6.1
pyOpenSSL==16.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycrypto==2.6.1
python-dateutil==2.5.2
python-openid==2.2.5
python-social-auth==0.2.14
requests==2.9.1
requests-oauthlib==0.6.1
rsa==3.4.2
simplejson==3.8.2
six==1.10.0
sure==1.2.7
uritemplate==0.6
urllib3==1.7.1
whitenoise==2.0.6
wsgiref==0.1.2
